requirements.txt
appdirs==1.4.3
asgiref==3.2.7
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
pipenv==2018.11.26
psycopg2==2.8.5
pytz==2020.1
selenium==3.141.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.20
virtualenv-clone==0.5.4
whitenoise==5.0.1
